+++ conf/system.php
@@ -17,12 +17,6 @@ return [
     'assets_source_path' => app_path() . DS . 'assets',
     'assets_path' => web_root() . DS . 'assets',
 
-    'environment_by_host' => [
-        'localhost' => 'development',
-        '127\.0\.0\.1' => 'development',
-        '(www\.)?mydomain\.com' => 'production',
-    ],
-
     /*
      * Session configuration
      */